add date custom field
