+++ Makefile
@@ -79,7 +79,7 @@ dev:: ## Run the Sihl app, watch files and restart on change
 		PID=$$!
 	fi
 	echo
-	inotifywait -e modify -e move -e create -e delete -e attrib -r `pwd` --exclude "(_build|logs|Makefile|.vscode|.devcontainer|.git|.DS_Store|node_modules|resources)" -qq
+	inotifywait -e modify -e move -e create -e delete -e attrib -r `pwd` --exclude "(_build|public|logs|Makefile|.vscode|.devcontainer|.git|.DS_Store|node_modules|resources)" -qq
 	[[ $$(jobs -pr) = "" ]] || kill $$(jobs -pr)
 	echo
 	done
