+++ pool/app/filter/repo/repo_utils.ml
@@ -74,7 +74,12 @@ let add_value_to_params operator value dyn =
     Ok
       (match value with
        | Bool b -> add Caqti_type.string (Utils.Bool.to_string b)
-       | Date d -> add Caqti_type.ptime d
+       | Date d ->
+         add Pool_common.Repo.Ptime.date d
+         (* TODO: Not sure if this works as expected:
+            https://stackoverflow.com/questions/2758486/mysql-compare-date-string-with-string-from-datetime-field
+
+            https://www.tutorialspoint.com/compare-date-strings-in-mysql*)
        | Language lang -> add Caqti_type.string (Pool_common.Language.show lang)
        | Nr n -> add Caqti_type.float n
        | Option id ->
