+++ pool/app/pool_common/utils_time.ml
@@ -20,3 +20,21 @@ let parse_time_span str =
     |> CCOption.to_result error
     >|= fun h -> h *. 3600. |> CCInt.of_float |> Ptime.Span.of_int_s
 ;;
+
+let parse_date str =
+  let open CCResult in
+  let error = Entity_message.(Invalid Field.Date) in
+  let split_date_string date =
+    date
+    |> CCString.split_on_char '-'
+    |> CCList.map CCInt.of_string
+    |> CCOption.sequence_l
+    |> CCOption.to_result error
+  in
+  str
+  |> split_date_string
+  >>= function
+  | [ y; m; d ] ->
+    (y, m, d) |> Ptime.of_date |> CCOption.to_result error >|= Ptime.to_date
+  | _ -> Error error
+;;
