+++ resources/admin.js
@@ -1,9 +1,7 @@
 import { initRoleSearchForm } from "./admin/role-search.js"
 import { initFilterForm } from "./admin/filter.js"
 import { initRichTextEditor } from "./admin/richTextEditor.js"
-import { initDatepicker } from "./admin/flatpickr.js"
 
-initDatepicker();
 initRoleSearchForm();
 initFilterForm();
 initRichTextEditor();
