+++ resources/index.js
@@ -1,3 +1,6 @@
 import 'htmx.org'
 import './index.scss'
+import { initDatepicker } from "./flatpickr.js"
 import framework from '../node_modules/@econ/frontend-framework/dist/main'
+
+initDatepicker();
